+++ route_link.c
@@ -488,7 +488,7 @@ static int append_link_state(struct overlay_buffer *payload, char flags,
     flags|=FLAG_HAS_ACK;
   if (drop_rate!=-1)
     flags|=FLAG_HAS_DROP_RATE;
-  int length_pos = ob_position(payload);
+  size_t length_pos = ob_position(payload);
   ob_append_byte(payload, 0);
   ob_append_byte(payload, flags);
   overlay_address_append(NULL, payload, receiver);
@@ -507,7 +507,7 @@ static int append_link_state(struct overlay_buffer *payload, char flags,
   if (ob_overrun(payload))
     return -1;
   // patch the record length
-  int end_pos = ob_position(payload);
+  size_t end_pos = ob_position(payload);
   ob_set(payload, length_pos, end_pos - length_pos);
   ob_checkpoint(payload);
   return 0;
@@ -898,7 +898,7 @@ static void link_send(struct sched_ent *alarm)
     ob_limitsize(frame->payload, 400);
     overlay_mdp_encode_ports(frame->payload, MDP_PORT_LINKSTATE, MDP_PORT_LINKSTATE);
     ob_checkpoint(frame->payload);
-    int pos = ob_position(frame->payload);
+    size_t pos = ob_position(frame->payload);
     enum_subscribers(NULL, append_link, frame->payload);
     ob_rewind(frame->payload);
     if (ob_position(frame->payload) == pos)
@@ -1243,7 +1243,7 @@ int link_receive(struct overlay_frame *frame, overlay_mdp_frame *mdp)
 
     struct subscriber *receiver=NULL, *transmitter=NULL;
     struct overlay_interface *interface = NULL;
-    int start_pos = ob_position(payload);
+    size_t start_pos = ob_position(payload);
     int length = ob_get(payload);
     if (length <=0)
       break;
